Migrate rest_tornado saltnado tests to pytest
Port the AsyncTestCase suite in
tests/unit/netapi/rest_tornado/test_saltnado.py to pytest, driving the
handler coroutines with the io_loop fixture, and remove the legacy file.
